fix(sdk): round-2 review — unsigned risk score, atomic risk+level, init cleanup
- Risk score crossed the JNI boundary as u32->jint and wrapped negative for
  saturating scores (>= 2^31). nativeEvaluateRisk now returns a LongArray and
  CoreRiskScore.score / RiskAssessment.score are Long, holding the full
  0..u32::MAX range (matching iOS UInt32). (BUG)
- evaluateRisk()+computeRiskLevel() were two independently-locked FFI calls; a
  config swap could land between them and mix policies. Added
  evaluateRiskAndLevel() that runs both under one read-lock acquisition (Android
  reentrant read lock; iOS single concurrent DispatchQueue.sync) and the SDK
  now uses it. (ANALYSIS)
- KsealSDK.initialize (Android) now frees the freshly-created core if a later
  init step throws, instead of leaking the native handle. iOS already releases
  it via ARC. (ANALYSIS)
- Clarified the Clock doc: it is a wall-clock seam (intentional for calendar
  hour bucketing), not monotonic. (ANALYSIS)

Test literals updated from Int to Long (20L/0L) as a mechanical consequence of
the score type widening; assertions are otherwise unchanged.

+++ sdk/android/src/main/kotlin/io/kseal/sdk/Model.kt
@@ -68,7 +68,7 @@ enum class Platform(val code: Int) {
 data class RiskAssessment(
     val riskBits: Long,
     val signals: Set<RiskSignal>,
-    val score: Int,
+    val score: Long,
     val confidence: Confidence,
     val trustLevel: TrustLevel,
 ) {
